usr/Jobs/r18e.batcher/requirements.txt
# requirements.txt
surrealdb
transformers
torch
torchvision
torchaudio
Pillow
accelerate
obstore